Separate namespace.ts Use sectoreOne time
