+++ front/src/models/api.ts
@@ -1,4 +1,5 @@
 import { Socket, io } from 'socket.io-client';
+import type { models } from '@/models/namespace';
 
 const routeApi: string = import.meta.env.VITE_ROUTE_API;
 
@@ -70,52 +71,6 @@ export class websocket {
     this.socket.on('updatedUserRaces', callback);
     return this;
   }
-}
-
-
-export namespace models {
-  export interface racesData {
-   races: [
-     {
-       id_race: number,
-       race_start: Date | string,
-       race_finish: Date | string,
-       id_car: number,
-       total_time: Date | string
-     }
-   ],
-    rank: number
-  }
-
-  export interface Avatar {
-    bgColor: string,
-    hatColor:  string,
-    faceColor:  string,
-    hairColor:  string,
-    shirtColor:  string,
-    hairColorRandom: boolean,
-    sex: string,
-    earSize: string,
-    hatType: string,
-    eyeType: string,
-    hairType: string,
-    noseType: string,
-    mouthType: string,
-    shirtType: string,
-    eyeBrowType: string,
-    glassesType: string,
-    shape: string
-  }
-
-  export type rankingData = {
-    id_race: number,
-    car: {
-      id_car: number,
-      pseudo: string,
-      avatar: Avatar,
-    },
-    total_time: Date | string
-  }[]
 }
 
 export default restful;
